requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.5
django-crispy-forms==1.9.1
django-extensions==2.2.9
django-froala-editor==3.1.0
django-js-asset==1.2.2
django-storages==1.9.1
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.4
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
